feat: add db rule
