+++ index.js
@@ -1,7 +1,8 @@
 module.exports = () => ({
     ports: [],
     modules: {
-        rule: require('./api/script')
+        rule: require('./api/script'),
+        'db/rule': require('./api/sql')
     },
     validations: {
         rule: require('./validations')
